Corrections regarding exclusions of small classes
In the interest of avoiding the vagaries of using GLP% for small classes, exclusions have been made to the disag data frames. This is explicit in 1.Rmd in the "Read Disag Files" section. Classes are defined by the triplet (school_code, subject, grade). Joins are made to eliminate classes when num_tested < 20 for subgroup=="ALL".